Implemented packet timeout
Added an option to discard a partially parsed packet if no data was
received for a while.  This timeout is expressed in server cycles.
Tests have shown that LPC11U37 microprocessor running at 48MHz gives 1
second timeout when the timeout count is set to 300000. However timeout
depends on many factors therefore this number should not be taken for
granted.